Refactor node widgets
